requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
torch>=2.0.0
huggingface-hub>=0.20.0
faiss-cpu>=1.8.0
groq>=0.4.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langgraph>=0.0.20
python-dotenv>=1.0.0
tqdm>=4.65.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
